requirements.txt
aws-cdk-lib==2.232.1
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v30
aws-cdk.aws-redshift-alpha
cdk-nag
boto3